requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.5
alembic==1.13.1

# Celery
celery==5.3.6
redis==5.0.1
flower==2.0.1

# AI/ML
groq==0.30.0
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
sentence-transformers==2.3.1
torch==2.10.0

# OCR & Document Processing
paddlepaddle==2.6.2
paddleocr==2.7.3
pdf2image==1.17.0
Pillow==10.2.0
pymupdf==1.23.22

# Vietnamese NLP
underthesea==6.8.4

# Search
rank-bm25==0.2.2
numpy==1.26.4

# Utilities
gdown==5.1.0
python-dotenv==1.0.1
httpx==0.26.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
